+++ R/class-PKNCAdose.R
@@ -248,6 +248,12 @@ getGroups.PKNCAdose <- function(...) {
   getGroups.PKNCAconc(...)
 }
 
+#' @describeIn group_vars.PKNCAconc Get group_vars for a PKNCAdose object
+#' @exportS3Method dplyr::group_vars
+group_vars.PKNCAdose <- function(x) {
+  group_vars.PKNCAconc(x)
+}
+
 #' @rdname getData.PKNCAconc
 #' @export
 #' @importFrom nlme getData
@@ -299,7 +305,3 @@ print.PKNCAdose <- function(x, n=6, summarize=FALSE, ...) {
 #' @rdname print.PKNCAconc
 #' @export
 summary.PKNCAdose <- summary.PKNCAconc
-
-#' @rdname split.PKNCAconc
-#' @export
-split.PKNCAdose <- split.PKNCAconc